requirements.txt
fastapi==0.64.0
uvicorn==0.14.0
pillow==8.3.1
python-multipart==0.0.5
aiofiles==0.7.0
jinja2==3.0.1